Split check file into global and no freeze files
